Enhance caption extraction and rate limiter precision
CaptionAwareProcessor now stores both the full caption line and the extracted content, improving downstream processing. The RateLimiter's burst window is now a float for finer control, and related tests have been updated to reflect these changes and ensure more robust coverage.

+++ src/utils/rate_limiter.py
@@ -41,7 +41,7 @@ class RateLimiter:
         requests_per_minute: int = 60,
         requests_per_hour: int = 1000,
         burst_limit: int = 10,
-        burst_window_seconds: int = 1,
+        burst_window_seconds: float = 1.0,
     ):
         """
         Initialize rate limiter.
